Alterações.

+++ ClassLib/Conta.cs
@@ -3,7 +3,7 @@ namespace conta
   public abstract class Conta
   {
     public virtual string Id { get; set; }
-    public decimal saldo { get; set; }
+    public double saldo { get; set; }
     public string titular { get; set; }
 
     
@@ -12,11 +12,11 @@ namespace conta
       this.titular = titular;
     }
 
-    public virtual void Depositar(decimal valor)
+    public virtual void Depositar(double valor)
     {
       this.saldo += valor;
     }
-    public virtual void Sacar(decimal valor)
+    public virtual void Sacar(double valor)
     {
       this.saldo -= ( valor <= saldo ? valor : 0);
     }
